Revert "Rename lib.py for my_lib.py"
This reverts commit 97141dc6c071caf544ba2076d31b7fa95056ac98.
