+++ python/bin.py
@@ -1,4 +1,4 @@
-from my_lib import sum
+from lib import sum
 
 if __name__ == "__main__":
     print(f"Answer of life 2 + 40 = {sum(2,40)}")
